pyproject.toml
[tool.poetry]
name = "camb-ai-pip"
version = "0.1.0"
description = ""
authors = ["Chandrashekhar R <hi.cskhere@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.32.2"
vcrpy = "^6.0.1"
pytest = "^8.2.1"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
camb-ai-pip = "camb_ai_pip.camb_client:run_test"
